chore: add sense build
