+++ package.json
@@ -29,6 +29,7 @@
   "module": "dist/sn-org-chart.esm.js",
   "scripts": {
     "build": "nebula build && shx cp assets/* dist",
+    "sense": "nebula sense --ext ./src/extension/ext-raw.js --meta ./src/extension/meta.json",
     "lint": "eslint src",
     "start": "nebula serve --type sn-org-chart",
     "test:unit": "aw -c aw.config.js",
