+++ src/extension/ext.js
@@ -1,0 +1,5 @@
+import extDef from './ext-raw';
+
+export default function ext(/* env */) {
+  return extDef;
+}
